backend/pyproject.toml
[project]
name = "porubly-api"
version = "0.1.0"
description = "REST API for По Рублю charity platform"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "gunicorn>=23",
    "sqlalchemy[asyncio]>=2.0",
    "asyncpg>=0.30",
    "alembic>=1.14",
    "pydantic[email]>=2.10",
    "pydantic-settings>=2.6",
    "redis>=5.2",
    "structlog>=24",
    "slowapi>=0.1",
    "httpx>=0.28",
    "PyJWT>=2.10",
    "cryptography>=44",
    "argon2-cffi>=23",
    "uuid-utils>=0.9",
    "python-multipart>=0.0.18",
    "boto3>=1.35",
    "taskiq[reload]>=0.11",
    "taskiq-redis>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8",
    "pytest-asyncio>=0.24",
    "ruff>=0.8",
    "httpx>=0.28",
]

[tool.ruff]
target-version = "py311"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "session"
asyncio_default_test_loop_scope = "session"
testpaths = ["tests"]
filterwarnings = [
    "ignore::DeprecationWarning:argon2",
]
